feat: persistent best brain across training sessions + u128 score + Up key
- Add GeneticAlgorithm::from_brain() to seed population from saved brain,
  so restarting training no longer resets to random.
- Change agent score type from u32 to u128 to prevent overflow.
- Fix clippy absurd_extreme_comparisons: use == u128::MAX instead of >=.
- Add Up arrow key as jump input alongside Space.

+++ src/ai/agent.rs
@@ -28,7 +28,7 @@ pub struct WorldAgent {
     pub on_ground: bool,
     pub alive: bool,
     pub network: DinoNet<B>,
-    pub score: u32,
+    pub score: u128,
     pub score_timer: f32,
     pub fitness: f32,
     pub ground_y: f32,
